pyproject.toml
[project]
name = "job-research-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "evjohns005",email = "ev.johns005@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry.dependencies]
pocketflow= ">=0.0.3,<0.0.4"
pypdf = "^6.9.1"
google-genai = "^1.68.0"
dotenv = "^0.9.9"

[tool.poetry]
packages = [{include = "job-research-agent"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
